Don't download stemcell if it already exists
This affects install-pcf/gcp and install-pcf/vsphere.

+++ tasks/upload-product-and-stemcell/task.sh
@@ -4,24 +4,28 @@ if [[ ! -z "$NO_PROXY" ]]; then
   echo "$OM_IP $OPS_MGR_HOST" >> /etc/hosts
 fi
 
-STEMCELL_VERSION=`cat ./pivnet-product/metadata.json | jq '.Dependencies[] | select(.Release.Product.Name | contains("Stemcells")) | .Release.Version'`
+STEMCELL_VERSION=`cat ./pivnet-product/metadata.json | jq --raw-output '.Dependencies[] | select(.Release.Product.Name | contains("Stemcells")) | .Release.Version'`
 
 if [ -n "$STEMCELL_VERSION" ]; then
-  echo "Downloading stemcell $STEMCELL_VERSION"
-  pivnet-cli login --api-token="$PIVNET_API_TOKEN"
-  pivnet-cli download-product-files -p stemcells -r $STEMCELL_VERSION -g $STEMCELL_GLOB --accept-eula
+  stemcell_exists=$(om-linux -t https://$OPS_MGR_HOST -u $OPS_MGR_USR -p $OPS_MGR_PWD -k curl --silent --path "/api/v0/diagnostic_report" | jq '.stemcells' | grep $STEMCELL_VERSION | grep -c ${STEMCELL_GLOB//\*/})
 
-  SC_FILE_PATH=`find ./ -name *.tgz`
+  if [[ "$stemcell_exists" -eq "0" ]]; then
+    echo "Downloading stemcell $STEMCELL_VERSION"
+    pivnet-cli login --api-token="$PIVNET_API_TOKEN"
+    pivnet-cli download-product-files -p stemcells -r $STEMCELL_VERSION -g $STEMCELL_GLOB --accept-eula
 
-  if [ ! -f "$SC_FILE_PATH" ]; then
-    echo "Stemcell file not found!"
-    exit 1
+    SC_FILE_PATH=`find ./ -name *.tgz`
+
+    if [ ! -f "$SC_FILE_PATH" ]; then
+      echo "Stemcell file not found!"
+      exit 1
+    fi
+
+    om-linux -t https://$OPS_MGR_HOST -u $OPS_MGR_USR -p $OPS_MGR_PWD -k upload-stemcell -s $SC_FILE_PATH
+
+    echo "Removing downloaded stemcell $STEMCELL_VERSION"
+    rm $SC_FILE_PATH
   fi
-
-  om-linux -t https://$OPS_MGR_HOST -u $OPS_MGR_USR -p $OPS_MGR_PWD -k upload-stemcell -s $SC_FILE_PATH
-
-  echo "Removing downloaded stemcell $STEMCELL_VERSION"
-  rm $SC_FILE_PATH
 fi
 
 FILE_PATH=`find ./pivnet-product -name *.pivotal`
